+++ src/main/java/bsf/digital/bsffinance/model/TransferResponseEntity.java
@@ -2,13 +2,14 @@ package bsf.digital.bsffinance.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+
 public class TransferResponseEntity {
     @JsonProperty("creditAccount")
     private Account creditAccount;
     @JsonProperty("debitAccount")
     private Account debitAccount;
     @JsonProperty("amount")
-    private Long amount;
+    private Float amount;
 
     public Account getCreditAccount() {
         return creditAccount;
@@ -26,11 +27,11 @@ public class TransferResponseEntity {
         this.debitAccount = debitAccount;
     }
 
-    public Long getAmount() {
+    public Float getAmount() {
         return amount;
     }
 
-    public void setAmount(Long amount) {
+    public void setAmount(Float amount) {
         this.amount = amount;
     }
 
@@ -42,7 +43,7 @@ public class TransferResponseEntity {
         this.creditAccount=creditAccount;
         return this;
     }
-    public TransferResponseEntity amount(Long amount){
+    public TransferResponseEntity amount(Float amount){
         this.amount =amount;
         return this;
     }
